move readOutlinedButtonStyle() to IntUiBridgeButton
